+++ ui/redux/rootReducer.js
@@ -18,8 +18,6 @@ import {
   RECEIVE_SAVED_SEARCHES,
   REQUEST_SAVED_SEARCHES,
   REQUEST_SAVED_VARIANTS,
-  REQUEST_SEARCHED_VARIANTS,
-  RECEIVE_SEARCHED_VARIANTS,
   REQUEST_PROJECT_DETAILS,
   REQUEST_ANALYSIS_GROUPS,
   RECEIVE_ANALYSIS_GROUPS,
@@ -273,8 +271,6 @@ const rootReducer = combineReducers({
   variantFunctionalDataByGuid: createObjectsByIdReducer(RECEIVE_DATA, 'variantFunctionalDataByGuid'),
   searchesByHash: createObjectsByIdReducer(RECEIVE_SAVED_SEARCHES, 'searchesByHash'),
   searchFamiliesByHash: createObjectsByIdReducer(RECEIVE_SAVED_SEARCHES, 'searchFamiliesByHash'),
-  searchedVariants: createSingleValueReducer(RECEIVE_SEARCHED_VARIANTS, []),
-  searchedVariantsLoading: loadingReducer(REQUEST_SEARCHED_VARIANTS, RECEIVE_SEARCHED_VARIANTS),
   savedSearchesByGuid: createObjectsByIdReducer(RECEIVE_SAVED_SEARCHES, 'savedSearchesByGuid'),
   savedSearchesLoading: loadingReducer(REQUEST_SAVED_SEARCHES, RECEIVE_SAVED_SEARCHES),
   transcriptsById: createObjectsByIdReducer(RECEIVE_DATA, 'transcriptsById'),
