+++ ui/redux/selectors.js
@@ -43,11 +43,9 @@ export const getWarningMessages = state => state.meta.warningMessages
 export const getAnvilLoadingDelayDate = state => state.meta.anvilLoadingDelayDate
 export const getSavedVariantsIsLoading = state => state.savedVariantsLoading.isLoading
 export const getSavedVariantsLoadingError = state => state.savedVariantsLoading.errorMessage
+// TODO
 export const getSearchesByHash = state => state.searchesByHash
 export const getSearchFamiliesByHash = state => state.searchFamiliesByHash
-export const getSearchedVariants = state => state.searchedVariants
-export const getSearchedVariantsIsLoading = state => state.searchedVariantsLoading.isLoading
-export const getSearchedVariantsErrorMessage = state => state.searchedVariantsLoading.errorMessage
 
 const groupEntitiesByProjectGuid = entities => Object.entries(entities).reduce((acc, [entityGuid, entity]) => {
   if (!(entity.projectGuid in acc)) {
@@ -168,6 +166,7 @@ export const getIGVSamplesByFamilySampleIndividual = createSelector(
   }, {}),
 )
 
+//  TODO
 export const getProjectDatasetTypes = createSelector(
   getProjectsByGuid,
   getSamplesGroupedByProjectGuid,
